Create BedMNasePEToWigNucleosomeMids.py
This script converts from paired-end MNase data in bed format to a format suitable for the BBG lab's nucleoosme calling pipeline. This is accomplished by first converting the paired-end bed data to unpaired nucleosome midpoints, which are assumed to be 73 bp 3' from the 5' end of each entry. Next, the resulting data is sorted and converted to fixed-step wig format.