Create FindCommonLoci.py
This script takes multiple bed files which have been annotated by the AnnotatePeaks script to include information on any genomic loci the entries are encompassed by. These data are used to count how many files each locus is present it. Then, information on the loci, their descriptions, and the number of files they are found in is printed to an output tsv file.